triehash is separated from util
