+++ ethcore/Cargo.toml
@@ -61,6 +61,7 @@ bloomable = { path = "../util/bloomable" }
 vm = { path = "vm" }
 wasm = { path = "wasm" }
 hash = { path = "../util/hash" }
+triehash = { path = "../util/triehash" }
 
 [dev-dependencies]
 native-contracts = { path = "native_contracts", features = ["test_contracts"] }
