+++ ethcore/light/Cargo.toml
@@ -33,6 +33,7 @@ serde = "1.0"
 serde_derive = "1.0"
 stats = { path = "../../util/stats" }
 hash = { path = "../../util/hash" }
+triehash = { path = "../../util/triehash" }
 
 [features]
 default = []
