+++ ethcore/src/snapshot/block.rs
@@ -23,7 +23,7 @@ use hash::keccak;
 use views::BlockView;
 use rlp::{DecoderError, RlpStream, UntrustedRlp};
 use util::{Bytes, H256};
-use util::triehash::ordered_trie_root;
+use triehash::ordered_trie_root;
 
 const HEADER_FIELDS: usize = 8;
 const BLOCK_FIELDS: usize = 2;
@@ -190,7 +190,7 @@ mod tests {
 		b.transactions.push(t2.into());
 
 		let receipts_root = b.header.receipts_root().clone();
-		b.header.set_transactions_root(::util::triehash::ordered_trie_root(
+		b.header.set_transactions_root(::triehash::ordered_trie_root(
 			b.transactions.iter().map(::rlp::encode).map(|out| out.into_vec())
 		));
