+++ ethcore/src/snapshot/consensus/work.rs
@@ -222,7 +222,7 @@ impl Rebuilder for PowRebuilder {
 		use views::BlockView;
 		use snapshot::verify_old_block;
 		use util::U256;
-		use util::triehash::ordered_trie_root;
+		use triehash::ordered_trie_root;
 
 		let rlp = UntrustedRlp::new(chunk);
 		let item_count = rlp.item_count()?;
