+++ sync/Cargo.toml
@@ -19,6 +19,7 @@ ethcore-light = { path = "../ethcore/light"}
 ethcore = { path = "../ethcore" }
 rlp = { path = "../util/rlp" }
 hash = { path = "../util/hash" }
+triehash = { path = "../util/triehash" }
 clippy = { version = "0.0.103", optional = true}
 log = "0.3"
 env_logger = "0.4"
