+++ util/src/lib.rs
@@ -117,14 +117,12 @@ pub mod common;
 pub mod error;
 pub mod bytes;
 pub mod misc;
-pub mod vector;
 pub mod hashdb;
 pub mod memorydb;
 pub mod migration;
 pub mod overlaydb;
 pub mod journaldb;
 pub mod kvdb;
-pub mod triehash;
 pub mod trie;
 pub mod nibbleslice;
 pub mod nibblevec;
@@ -138,14 +136,12 @@ pub use hashdb::*;
 pub use memorydb::MemoryDB;
 pub use overlaydb::*;
 pub use journaldb::JournalDB;
-pub use triehash::*;
 pub use trie::{Trie, TrieMut, TrieDB, TrieDBMut, TrieFactory, TrieError, SecTrieDB, SecTrieDBMut};
 pub use semantic_version::*;
 pub use kvdb::*;
 pub use timer::*;
 pub use error::*;
 pub use bytes::*;
-pub use vector::*;
 pub use bigint::prelude::*;
 pub use bigint::hash;
